initial window creation
Note: I would've liked to keep the depencies as low as possible and now
there's 108 of them..
I'm not too worried about the transitive depencies and it just makes the
crate count go up, but I would've liked to implement a logger like
env_logger myself. It might've not been a lot of work or very
interesting but I guess that's one less thing to worry about now.